Fix snackbar alert for change password
